refactor: implement chat message processing in a dedicated function
- Moved chat message processing logic from the API route to a new `processChatMessage` function in `chatProcessor.ts`.
- Added input validation for the message in the API route.
- Updated the API route to handle chat message processing and response generation using the new function.
- Ensured error handling and logging are maintained during the processing of chat messages.
